+++ general_utilities/job_management/subjob_test.py
@@ -4,6 +4,7 @@ import dxpy
 import gzip
 from pathlib import Path
 from general_utilities.job_management.subjob_utility import SubjobUtility
+# from general_utilities.job_management.subjob_subpackage.subjob_test import tabix_subjob
 from general_utilities.association_resources import download_dxfile_by_name, generate_linked_dx_file
 from general_utilities.mrc_logger import MRCLogger
 
@@ -57,27 +58,3 @@ def test_subjob(tabix_dxfile: dxpy.DXFile):
                     output_files.append(subjob_output)
 
     return output_files
-
-
-@dxpy.entry_point('tabix_subjob')
-def tabix_subjob(input_table: dict, chromosome: str):
-    local_tab = download_dxfile_by_name(input_table, print_status=True)
-    output_tab = Path(f'chr{chromosome}.tsv')
-
-    with local_tab.open('r') as local_open,\
-        output_tab.open('w') as local_out:
-
-        local_csv = csv.DictReader(local_open, delimiter='\t')
-        output_csv = csv.DictWriter(local_out, delimiter='\t', fieldnames=local_csv.fieldnames)
-
-        output_csv.writeheader()
-
-        match_string = f'chr{chromosome}'
-
-        for row in local_csv:
-            if row['#CHROM'] == match_string:
-                output_csv.writerow(row)
-
-    output = {'chromosome': chromosome, 'subset_tsv': generate_linked_dx_file(output_tab)}
-    return output
-
